chore: update plugin and support setup
