+++ cypress/support/e2e.js
@@ -3,4 +3,3 @@ import { enableAutoLogin } from '@dhis2/cypress-commands'
 import './commands.js'
 
 enableAutoLogin()
-// enableNetworkShim()
